Add logging stacktrace for go-pkgs errors
