+++ go.mod
@@ -16,12 +16,12 @@ require (
 	github.com/spf13/pflag v1.0.3 // indirect
 	github.com/streadway/amqp v0.0.0-20200108173154-1c71cc93ed71
 	github.com/stretchr/objx v0.2.0 // indirect
-	github.com/stretchr/testify v1.4.0
-	go.uber.org/zap v1.13.0
+	github.com/stretchr/testify v1.7.0
+	github.com/sumup-oss/go-pkgs/errors v0.0.0-20210806131309-dd59fc2fd123
+	go.uber.org/zap v1.19.0
 	golang.org/x/crypto v0.0.0-20190510104115-cbcb75029529
 	golang.org/x/sys v0.0.0-20200122134326-e047566fdf82 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	gopkg.in/yaml.v2 v2.2.5 // indirect
 )
 
 go 1.13
