+++ logger/http_fields.go
@@ -1,4 +1,4 @@
-// Copyright 2019 SumUp Ltd.
+// Copyright 2021 SumUp Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -22,27 +22,6 @@ import (
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
 )
-
-// ErrorField creates a zap.Field for the corresponding error.
-//
-// The sole purpose of this func is to provide extensibility, on the way errors logging is handled.
-// Currently it does not do anything different than zap.Error().
-//
-// Future version should wrap the error with an object that supports Errors interface:
-//   type Errors interface {
-//     Errors() []error
-//   }
-func ErrorField(err error) zap.Field {
-	if err == nil {
-		return zap.Skip()
-	}
-
-	return zap.Field{
-		Key:       "error",
-		Type:      zapcore.ErrorType,
-		Interface: err,
-	}
-}
 
 type httpRequestDumpField struct {
 	req      *http.Request
